Give proper format and naming to the library
