+++ src/MS5637.cpp
@@ -5,7 +5,7 @@
  *
  * Written by Angus Gratton (angus at freetronics dot com)
  */
-#include "BaroSensor.h"
+#include "MS5637.h"
 
 /* i2c address of module */
 #define BARO_ADDR 0x76
@@ -27,11 +27,11 @@ const uint8_t SamplingDelayMs[6] PROGMEM = {
 #define CMD_START_D2(oversample_level) (0x50 + 2*(int)oversample_level)
 #define CMD_READ_ADC 0x00
 
-BaroSensorClass BaroSensor;
+MS5637Class MS5637;
 
 // Temporary hack due to bug in Arduino 1.5.0-1.5.6 on ARM (see below)
-inline static int8_t _endTransmission(bool stop = true)
-{
+inline static int8_t _endTransmission(bool stop = true) {
+  
   int8_t res = Wire.endTransmission(stop);
 #ifdef __AVR__
   return res;
@@ -41,9 +41,10 @@ inline static int8_t _endTransmission(bool stop = true)
   // This is corrected in https://github.com/arduino/Arduino/pull/1994 but not yet released.
   return 0;
 #endif
+
 }
 
-void BaroSensorClass::begin()
+void MS5637Class::begin()
 {
   Wire.begin();
   Wire.beginTransmission(BARO_ADDR);
@@ -78,7 +79,7 @@ void BaroSensorClass::begin()
   initialised = true;
 }
 
-float BaroSensorClass::getTemperature(TempUnit scale, BaroOversampleLevel level)
+float MS5637Class::getTemperature(TempUnit scale, BaroOversampleLevel level)
 {
   float result;
   if(getTempAndPressure(&result, NULL, scale, level))
@@ -87,7 +88,7 @@ float BaroSensorClass::getTemperature(TempUnit scale, BaroOversampleLevel level)
     return NAN;
 }
 
-float BaroSensorClass::getPressure(BaroOversampleLevel level)
+float MS5637Class::getPressure(BaroOversampleLevel level)
 {
   float result;
   if(getTempAndPressure(NULL, &result, CELSIUS, level))
@@ -96,7 +97,7 @@ float BaroSensorClass::getPressure(BaroOversampleLevel level)
     return NAN;
 }
 
-bool BaroSensorClass::getTempAndPressure(float *temperature, float *pressure, TempUnit tempScale, BaroOversampleLevel level)
+bool MS5637Class::getTempAndPressure(float *temperature, float *pressure, TempUnit tempScale, BaroOversampleLevel level)
 {
   if(err || !initialised)
     return false;
@@ -156,7 +157,7 @@ bool BaroSensorClass::getTempAndPressure(float *temperature, float *pressure, Te
   return true;
 }
 
-uint32_t BaroSensorClass::takeReading(uint8_t trigger_cmd, BaroOversampleLevel oversample_level)
+uint32_t MS5637Class::takeReading(uint8_t trigger_cmd, BaroOversampleLevel oversample_level)
 {
   Wire.beginTransmission(BARO_ADDR);
   Wire.write(trigger_cmd);
@@ -185,7 +186,7 @@ uint32_t BaroSensorClass::takeReading(uint8_t trigger_cmd, BaroOversampleLevel o
   return result;
 }
 
-void BaroSensorClass::dumpDebugOutput()
+void MS5637Class::dumpDebugOutput()
 {
   Serial.print(F("C1 = 0x"));
   Serial.println(c1, HEX);
